+++ test/parquet_test.rb
@@ -83,6 +83,14 @@ class ParquetTest < Minitest::Test
     assert_equal ({"a" => Polars::Int64, "b" => Polars::String}), schema
   end
 
+  def test_read_parquet_metadata
+    metadata = Polars.read_parquet_metadata("test/support/data.parquet")
+    expected = {
+      "ARROW:schema" => "/////6gAAAAQAAAAAAAKAAwABgAFAAgACgAAAAABBAAMAAAACAAIAAAABAAIAAAABAAAAAIAAABAAAAABAAAANj///8AAAEFEAAAABgAAAAEAAAAAAAAAAEAAABiAAAABAAEAAQAAAAQABQACAAGAAcADAAAABAAEAAAAAAAAQIQAAAAHAAAAAQAAAAAAAAAAQAAAGEAAAAIAAwACAAHAAgAAAAAAAABQAAAAAAAAAA="
+    }
+    assert_equal expected, metadata
+  end
+
   def test_write_parquet
     df = Polars::DataFrame.new({"a" => [1, 2, 3], "b" => ["one", "two", "three"]})
     assert_nil df.write_parquet(temp_path)